Splitting channels.py into two files
